refactor: move repository pkg to internal
