+++ service/post_service.go
@@ -2,7 +2,7 @@ package service
 
 import (
 	"basic_server/internal/model"
-	"basic_server/repository"
+	"basic_server/internal/repository"
 	"net/http"
 )
